Made disk store use multiple directories, deleted ShuffleManager
